chore: move forecasts files
