+++ src/app/UI/admin/Forecasts/Forecast/problems/helpers/generateOptions.ts
